packages/server/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "memmachine-server"
dynamic = ["version"]
dependencies = [
    "memmachine-common",
    "alembic>=1.17.1",
    "asyncpg>=0.31.0",
    "aiosqlite>=0.20.0",
    "boto3>=1.40.40",
    "boto3-stubs==1.40.64",
    "cohere>=5.20.0",
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "fastmcp>=3.0.2",
    "instructor[bedrock]>=1.12.0",
    "json-repair>=0.54",
    "langchain-aws>=0.2.33",
    "neo4j>=5.28.2",
    "nltk>=3.9.3",
    "openai>=2.21.0",
    "pgvector>=0.4.1",
    "prometheus-client>=0.22.1",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "rank-bm25>=0.2.2",
    "regex>=2023.0.0",
    "sqlite-vec>=0.1.9",
    "sqlalchemy>=2.0.43",
    "usearch>=2.23.0",
    "uvicorn>=0.35.0",
    "greenlet>=3.2.4",
    "hnswlib>=0.8.0",
    "tzdata>=2024.1 ; sys_platform == 'win32'",
    "psutil>=7.1.0",
]
requires-python = ">= 3.12"
authors = [
  {name="MemMachine", email="noreply@memmachine.ai"},
]
description = "MemMachine Server - The complete MemMachine memory system server with episodic and profile memory"
readme = "README.md"
keywords = ["memmachine", "server", "memory", "ai", "agent"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Operating System :: OS Independent",
]
license = "Apache-2.0"

[project.optional-dependencies]
faiss = [
    "faiss-cpu>=1.13.0",
]
gpu = [
    "sentence-transformers>=5.1.0",
]
nebula = [
    "nebula5-python>=5.2.1",
]

[project.scripts]
memmachine-server = "memmachine_server.server.app:main"
memmachine-nltk-setup = "memmachine_server:setup_nltk"
memmachine-mcp-stdio = "memmachine_server.server.mcp_stdio:main"
memmachine-mcp-http = "memmachine_server.server.mcp_http:main"
memmachine-configure = "memmachine_server.installation.memmachine_configure:main"

[project.urls]
Homepage = "https://github.com/MemMachine/MemMachine"
Issues = "https://github.com/MemMachine/MemMachine/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.md", "*.txt", "*.yml", "*.yaml", "*.sql"]

[tool.setuptools_scm]
root = "../.."

[tool.uv]
package = true

[tool.ty.environment]
python-version = "3.12"
extra-paths = ["../common/src", "../client/src"]

[tool.ty.src]
exclude = ["build/**"]

[tool.ruff]
extend = "../../pyproject.toml"
target-version = "py312"
